Implementando restricao de conteudo: cliente so recupera ele mesmo

+++ src/main/java/com/alessandro/sistemaloja/dto/AuthenticatedUserDetails.java
@@ -2,5 +2,5 @@ package com.alessandro.sistemaloja.dto;
 
 import java.util.List;
 
-public record AuthenticatedUserDetails(Long id, String email, List<String> perfis) {
+public record AuthenticatedUserDetails(Integer id, String email, List<String> perfis) {
 }
